Create NotAOne
Not-A-One is a turn-based game where 2 players, user and computer, are give 3 random numbers . Both of them have to accumulate the point 150 or more  . However, the numbers cannot have a value 1. If one of the numbers has value 1 the player loses a turn and misses the opportunity to accumulate the point. If there are two 1s, the the player loses a turn and half of accumulated points. If the player receives three 1s, the player loses a turn and all of the points.