Create Cow&Bull
Cow and Bull is a turned based game where user receives a randomly generate, 4-digit number where all four digits are unique.The players have to guess a 4-digit number. For every digit that the digit the user guessed correctly in the correct place, they have a "bull" (as in bullseye). For every digit the user guessed correctly in the wrong place is a "cow". Every time the user makes a guess, the program tells how many "bulls" and "cows" they have. Once the user guesses the correct number, the game is over. The game also tracks
of the number of guesses the user makes throughout the game and the time it takes to correctly guess the 4-digit number.
At the end of the game tell the user how many guesses and how long it took to win.