Animal Quiz
In this game, a player is asked by computer to type in several animal names such as "dog," "cat," etc., in the beginning of the game session. When entering a return key without typing anything (i.e., empty line), this process of input cycle ends.   A word can be defined as a sequence of consecutive non-white-space characters with white space before and after.