My Ship Sails
My Ship Sails is a  kids card game. The aim of the game is to collect 7 cards of the "same suit" you choose. You can sail out your ship if you have 7 cards of the same suit. The card numbers are unimportant. You need to select a card to discard, and pass it to your neighboring player. Each player selects a discard at the same time. And passing is done in a clockwise order.